requirements.txt
line-bot-sdk
Flask
uWSGI==2.0.14
request
pytest
pytest-cov
coverage
pylint
pyflakes
flake8
pyyaml
jieba
gspread
oauth2client
cachetools
